+++ bot-components/GitHub_GraphQL.ml
@@ -261,6 +261,41 @@ module MergePullRequest =
   }
 |}]
 
+module ClosePullRequest =
+[%graphql
+{|
+  mutation closePullRequest($pr_id: ID!) {
+    closePullRequest(
+      input: {pullRequestId: $pr_id}) {
+      pullRequest {
+        state
+      }
+    }
+  }
+|}]
+
+module LabelPullRequest =
+[%graphql
+{|
+  mutation labelPullRequest($pr_id: ID!, $label_ids: [ID!]!) {
+    addLabelsToLabelable(
+      input: {labelableId: $pr_id, labelIds:$label_ids}) {
+      clientMutationId
+    }
+  }
+|}]
+
+module UnlabelPullRequest =
+[%graphql
+{|
+  mutation unlabelPullRequest($pr_id: ID!, $label_ids: [ID!]!) {
+    removeLabelsFromLabelable(
+      input: {labelableId: $pr_id, labelIds:$label_ids}) {
+      clientMutationId
+    }
+  }
+|}]
+
 module NewCheckRun =
 [%graphql
 {|
@@ -335,6 +370,94 @@ module GetCheckRuns =
       }
     }
   }
+|}]
+
+module GetLabel =
+[%graphql
+{|
+  query getLabels($owner: String!, $repo: String!, $label: String!) {
+    repository(owner:$owner, name:$repo) {
+      label(name: $label) {
+        id
+      }
+    }
+  }
+|}]
+
+module GetOpenPullRequestWithLabel =
+[%graphql
+{|
+
+query getOpenPullRequestWithLabel($owner: String!, $repo:String!, $label:String!, $cursor: String, $len: Int!) {
+  repository(name: $repo,owner:$owner) {
+    pullRequests(first: $len, labels: [$label], states: [OPEN], after: $cursor) {
+      nodes {
+        id
+        number
+      }
+      pageInfo {
+        endCursor
+        hasNextPage
+      }
+    }
+  }
+}
+
+|}]
+
+module GetPullRequestLabelTimeline =
+[%graphql
+{|
+
+query getPullRequestLabelTimeline($owner: String!, $repo:String!, $prNumber: Int!, $cursor: String, $len: Int!) {
+  repository(name: $repo,owner:$owner) {
+    pullRequest(number: $prNumber) {
+      timelineItems(itemTypes: [LABELED_EVENT, UNLABELED_EVENT], after: $cursor, first: $len) {
+        nodes {
+          ... on LabeledEvent {
+            createdAt
+            label {
+              name
+            }
+          }
+          ... on UnlabeledEvent {
+            createdAt
+            label {
+              name
+            }
+          }
+        }
+        pageInfo {
+          endCursor
+          hasNextPage
+        }
+      }
+    }
+  }
+}
+
+|}]
+
+module GetPullRequestLabels =
+[%graphql
+{|
+
+query getPullRequestLabels($owner: String!, $repo:String!, $prNumber: Int!, $cursor: String, $len: Int!) {
+  repository(name: $repo,owner:$owner) {
+    pullRequest(number: $prNumber) {
+      labels (after: $cursor, first: $len) {
+        nodes {
+          name
+        }
+        pageInfo {
+          endCursor
+          hasNextPage
+        }
+      }
+    }
+  }
+}
+
 |}]
 
 module GetBaseAndHeadChecks =