+++ bot-components/GitHub_mutations.mli
@@ -11,6 +11,8 @@ val post_comment :
 
 val report_on_posting_comment : (string, string) result -> unit Lwt.t
 
+val close_pull_request : bot_info:Bot_info.t -> pr_id:id -> unit Lwt.t
+
 val merge_pull_request :
      bot_info:Bot_info.t
   -> ?merge_method:merge_method
@@ -49,9 +51,10 @@ val update_check_run :
 val reflect_pull_request_milestone :
   bot_info:Bot_info.t -> issue_closer_info -> unit Lwt.t
 
-val add_rebase_label : bot_info:Bot_info.t -> issue -> unit Lwt.t
+val add_labels : bot_info:Bot_info.t -> labels:id list -> pr_id:id -> unit Lwt.t
 
-val remove_rebase_label : bot_info:Bot_info.t -> issue -> unit Lwt.t
+val remove_labels :
+  bot_info:Bot_info.t -> labels:id list -> pr_id:id -> unit Lwt.t
 
 val update_milestone : bot_info:Bot_info.t -> string -> issue -> unit Lwt.t