+++ src/actions.mli
@@ -85,3 +85,15 @@ val coq_bug_minimizer_resume_ci_minimization_action :
   -> app_id:int
   -> string
   -> (Cohttp.Response.t * Cohttp_lwt__Body.t) Lwt.t
+
+val coq_check_needs_rebase_pr :
+     bot_info:Bot_info.t
+  -> owner:string
+  -> repo:string
+  -> warn_after:int
+  -> close_after:int
+  -> throttle:int
+  -> unit Lwt.t
+
+val coq_check_stale_pr :
+  bot_info:Bot_info.t -> owner:string -> repo:string -> after:int -> unit Lwt.t